Add files via upload
Fixed a problem where spaces were being used as a delimiter instead of only commas. This was resulting in incorrect wmi queries when using strings that contained spaces. For example searching in logfiles that contain a space in the name. Like "DFS Replication". Details can be found in https://github.com/pulsejets/check_wmi_eventid/issues/4